Refactor Rendering Process
Fix vertical line; Add Renderer class
TODO: implement index arrays and rendering multiple objects

+++ Equivnox/src/Math/Matrix4x4.h
@@ -7,7 +7,7 @@ namespace EQX {
 	class Matrix4x4
 	{
 	public:
-		double mMat[4][4];
+		float mMat[4][4];
 		// TODO Matrix Operations
 		// TODO MVP Matrices
 	};
